Add Consumer Service for Managing Consumers and Plugins (#31)
* added an if condition to check for local environment and skip routes creation

* using local environment variable to skip checking routes for delete operation as well

* added gateway consumer service with endpoints to CRUDify consumer plugins

* added all the endpoints required for managing consumers

* formatted the code using autopep8

* added required transformations and validations

* updated endpoints with required permissions

* removed duplicate check for local environment and grouped oc login under single check

* removed all the endpoints except plugins and updated the permissions/scopes

* Updated validate tags and added verify tags to compare the request and plugin data

+++ microservices/gatewayApi/v2/routes/gateway.py
@@ -78,29 +78,30 @@ def delete_config(namespace: str, qualifier = "") -> object:
 
     elif cmd == "sync" and not local_environment:
         try:
-            route_count = prepare_apply_routes (namespace, selectTag, is_host_transform_enabled(), tempFolder)
-            log.debug("%s - Prepared %d routes" % (namespace, route_count))
-            if route_count > 0:
-                apply_routes (tempFolder)
-                log.debug("%s - Applied %d routes" % (namespace, route_count))
-            route_count = prepare_delete_routes (namespace, selectTag, tempFolder)
-            log.debug("%s - Prepared %d deletions" % (namespace, route_count))
-            if route_count > 0:
-                delete_routes (tempFolder)
-        
-            # create Network Security Policies (nsp) for any upstream that
-            # has the format: <name>.<ocp_ns>.svc
-            log.debug("%s - Update NSPs" % (namespace))
-            ocp_ns_list = get_ocp_service_namespaces (tempFolder)
-            for ocp_ns in ocp_ns_list:
-                if check_nsp (namespace, ocp_ns) is False:
-                    apply_nsp (namespace, ocp_ns, tempFolder)
+            if not local_environment:
+                route_count = prepare_apply_routes (namespace, selectTag, is_host_transform_enabled(), tempFolder)
+                log.debug("%s - Prepared %d routes" % (namespace, route_count))
+                if route_count > 0:
+                    apply_routes (tempFolder)
+                    log.debug("%s - Applied %d routes" % (namespace, route_count))
+                route_count = prepare_delete_routes (namespace, selectTag, tempFolder)
+                log.debug("%s - Prepared %d deletions" % (namespace, route_count))
+                if route_count > 0:
+                    delete_routes (tempFolder)
+            
+                # create Network Security Policies (nsp) for any upstream that
+                # has the format: <name>.<ocp_ns>.svc
+                log.debug("%s - Update NSPs" % (namespace))
+                ocp_ns_list = get_ocp_service_namespaces (tempFolder)
+                for ocp_ns in ocp_ns_list:
+                    if check_nsp (namespace, ocp_ns) is False:
+                        apply_nsp (namespace, ocp_ns, tempFolder)
 
-            # ok all looks good, so update a secret containing the original submitted request
-            log.debug("%s - Update Original Config" % (namespace))
-            write_submitted_config ("", tempFolder)
-            prep_and_apply_secret (namespace, selectTag, tempFolder)
-            log.debug("%s - Updated Original Config" % (namespace))
+                # ok all looks good, so update a secret containing the original submitted request
+                log.debug("%s - Update Original Config" % (namespace))
+                write_submitted_config ("", tempFolder)
+                prep_and_apply_secret (namespace, selectTag, tempFolder)
+                log.debug("%s - Updated Original Config" % (namespace))
         except HTTPException as ex:
             traceback.print_exc()
             log.error("Error updating custom routes, nsps and secrets. %s" % ex)
@@ -274,30 +275,32 @@ def write_config(namespace: str) -> object:
 
     elif cmd == "sync" and not local_environment:
         try:
-            route_count = prepare_apply_routes (namespace, selectTag, is_host_transform_enabled(), tempFolder)
-            log.debug("[%s] - Prepared %d routes" % (namespace, route_count))
-            if route_count > 0:
-                apply_routes (tempFolder)
-                log.debug("[%s] - Applied %d routes" % (namespace, route_count))
-            route_count = prepare_delete_routes (namespace, selectTag, tempFolder)
-            log.debug("[%s] - Prepared %d deletions" % (namespace, route_count))
-            if route_count > 0:
-                delete_routes (tempFolder)
-        
-            # create Network Security Policies (nsp) for any upstream that
-            # has the format: <name>.<ocp_ns>.svc
-            if should_we_apply_nsp_policies():
-                log.debug("[%s] - Update NSPs" % (namespace))
-                ocp_ns_list = get_ocp_service_namespaces (tempFolder)
-                for ocp_ns in ocp_ns_list:
-                    if check_nsp (namespace, ocp_ns) is False:
-                        apply_nsp (namespace, ocp_ns, tempFolder)
+            # skip creation of routes in local development environment
+            if not local_environment:
+                route_count = prepare_apply_routes (namespace, selectTag, is_host_transform_enabled(), tempFolder)
+                log.debug("[%s] - Prepared %d routes" % (namespace, route_count))
+                if route_count > 0:
+                    apply_routes (tempFolder)
+                    log.debug("[%s] - Applied %d routes" % (namespace, route_count))
+                route_count = prepare_delete_routes (namespace, selectTag, tempFolder)
+                log.debug("[%s] - Prepared %d deletions" % (namespace, route_count))
+                if route_count > 0:
+                    delete_routes (tempFolder)
+            
+                # create Network Security Policies (nsp) for any upstream that
+                # has the format: <name>.<ocp_ns>.svc
+                if should_we_apply_nsp_policies():
+                    log.debug("[%s] - Update NSPs" % (namespace))
+                    ocp_ns_list = get_ocp_service_namespaces (tempFolder)
+                    for ocp_ns in ocp_ns_list:
+                        if check_nsp (namespace, ocp_ns) is False:
+                            apply_nsp (namespace, ocp_ns, tempFolder)
 
-            # ok all looks good, so update a secret containing the original submitted request
-            log.debug("[%s] - Update Original Config" % (namespace))
-            write_submitted_config (orig_config, tempFolder)
-            prep_and_apply_secret (namespace, selectTag, tempFolder)
-            log.debug("[%s] - Updated Original Config" % (namespace))
+                # ok all looks good, so update a secret containing the original submitted request
+                log.debug("[%s] - Update Original Config" % (namespace))
+                write_submitted_config (orig_config, tempFolder)
+                prep_and_apply_secret (namespace, selectTag, tempFolder)
+                log.debug("[%s] - Updated Original Config" % (namespace))
         except HTTPException as ex:
             traceback.print_exc()
             log.error("[%s] Error updating custom routes, nsps and secrets. %s" % (namespace, ex))
